Control access to the s3 user data via instance profiles.
Previously this was done using a vpc endpoint with an s3 bucket policy
that allowed access to anything coming from that vpc. Of course, every
vpc we spun up broke the bucket policy for the others. So, we have
reversed the policy so it is controlled from the things doing the accessing
(ec2 instance profiles) rather than being controlled from the thing being
accessed (s3 bucket objects).