drain and delete bootstrap node when done
leaving the bootstrapper node around means that things like daemonsets
report back as NodeLost until everytully the cluster forgots about the
node. We can speed this up by draining and deleting the bootstraper node
after we are done with it.

--force and --ignore-daemonsets flags are used to make the drain more
agressive since this node is def going away, but may not always be required