Add lambda to optionally ship cloudwatch logs to splunk
- Decision was made to have one lambda per cluster rather than
one lambda per account. This is because the cloudwatch log
group to ship is set up with each cluster and is not known
by the account

- Default is not to ship logs to splunk for the moment as dev
stacks will not need to and EIDAS  only want to ship prod
logs rather than staging (although also want to ship staging
whilst the pen test is ongoing).

The code has been adapted from https://github.com/deliveryhero/dhh-system-engineering/tree/master/terraform/aws/modules/lambda_splunk_forwarder.
Whilst I tried to avoid diverging from the module, I ended up
bringing the code into our repository and adapting it. The reasons
for this are
- there is no way to optionally include a module as per https://github.com/hashicorp/terraform/issues/953
meaning that we can not turn the module on and off for dev/prod
stacks
- As we may be running multiple lambdas per AWS account they need
to have independant resource names so we make them unique by passing
the cluster name in as a variable
- Lambda names which contain the cluster name will be shown
as the source for splunk events, meaning we can differentiate
between clusters.
- Cyber security believe they will need to make changes to the
lambda file in the near future

Our lambdas create their own logs themselves which go into
Cloudwatch. I am not sure if this is of particular benefit but I
have kept this behaviour as it existed in `dhh-system-engineering`.