Add admins to DEV role by default
Right now the ignition breaks all of the clusters. This is caused by the
fact that the policy is not listing any principal users to be
applicable.

This is a developer readonly role, that currently is purely purposed
around interacting with Kubernetes API with read-only permissions.

In the future it will probably allow readonly access to resources on
AWS.

There is no reason for admin users not to have that ability on top of
their own.