Provide real fallback defaults for addons
When we implement a new feature right now, say add splunk into the mix,
we will break any other cluster setup that currently doesn't have that
key embeded.

Instead, we could be having defaults as a different map, and attempt to
merge them when necessary to avoid breaking the cluster.