Create ssh key so CI can write to github repos
Given that CI pipelines will exist in the default concourse
team (called `main`), k8s secrets that need to be available to
concourse must be placed in the `ci-system-main` namespace.
This is documented in the concourse chart:
https://github.com/helm/charts/tree/master/stable/concourse#kubernetes-secrets

This key is also outputted from the gsp-cluster module as it will
need to be manually copied by a developer and added as a
deployment key with read/write access to the repository which we
allow our CI to have write access to.

This will allow our CI to tag repositories as part of our
deployment pipeline, tagging a `staging` and `production` branch.