Use lambda written by cyber security team for splunk log forwarding
https://github.com/alphagov/cyber-cloudwatch-fluentd-to-hec

Cyber wrote their own lambda so they have better control
over the metadata that gets sent to splunk.

I have removed the lambda we were previously using and replaced it
with a zip of the new lambda. The zip was manually done by myself
of the master branch and copied to this repo and commited. This is
not an ideal solution and should be considered taking on some tech
debt but I think the trade off is worth it for the moment.

Other options I considered/tried that would mean there is not a
manual process for a developer to update this function if cyber
make changes to it.

- copy all the files from the cyber repo, commit them to this repo
 and zip them using terraform. This has exactly the same problem
with requiring a manual copy from a developer
- download the zip of the master branch as per https://stackoverflow.com/questions/51436429/how-can-i-use-a-git-repository-as-an-aws-lambda-source-within-terraform/51441406
But this when unzipped starts a top level folder of the repository
name, which is not a python module, and then the lambda can not
be run. I spent a little bit of time trying to edit the lambda
(adding __init__.py files for example and changing import paths)
but had no luck.

Given the small cost of manually zipping the file and commiting it,
plus that this will be a rare action I decided it's worth taking on
the small bit of tech debt for the moment.