Prevent sealed secret key rotation
We'd like to avoid the need of re-encrypting the secrets if say a
cluster had to be re-built.

This will make it harder to rotate that key for the cluster, as the code
will have to be commented out, applied and then destroyed, but it's a
necessary evil.

This also means that when any plan that includes a destroy of this
resource will return an error message, which is going to affect our
pipelines and perhaps cause a lot of confusion when the day comes.

It should however mean, that the actual key is going to persist in
terraform state file.