Allow disabling destroy prevention
Our sandbox accounts suffer from not being able to be torn down. We can
fix that by pointing to a different branch that will disable the
prevention.

If that's the case, we can provide the terraform with a variable, that
will set it for us. By default, prevent deletion, but if it's a sandbox
correctly configured, it should have an additional flag.