Workaround issue with calculating length of computed resources
We have to pass the number of subnets around (rather than dynamically
counting them) because of an issue in Terraform:

https://github.com/hashicorp/terraform/issues/12570

Without this we see the following error message when `terraform
apply`ing from scratch:

```
Error: Error refreshing state: 1 error(s) occurred:

* module.hsm.data.aws_subnet.vpc: data.aws_subnet.vpc: value of 'count'
cannot be computed
```

If the cluster has already been spun up and the HSM module included at a
later date then the Terraform would apply cleanly in its current state.

This may be end up being resolved in Terraform 0.12 when it is released.