Generate deploy key per release
GitHub has a rule, that there can only be one unique Deploy Key per
repository.

We're moving the terraform generation of these SSH keys to flux-release,
in order to accommodate that.

We're also generating a Secret to be consumed by the Concourse as well
as the config map to be easily extractable by the team and added to
their repository.