Add count for SSH key generation
We're using flux system everywhere wheather or not the release is
enabled.

This causes issues when for instance `ci-system` is disabled in one of
the clusters and still generates the key that it's confused what to do
with it.

It's a technical debt and should go away in the future.