Define owner role for pipeline-operator
We've been relying on backwards compatibility, which is sadly deprecated
in favour of this roles file. The owner/admin user is no longer assinged
to any team automatically, meaning we get a nice error message when we
attempt to use API with this auth.

It is difficult to find this page, so linking it up:

https://concourse-ci.org/managing-teams.html#setting-roles