pin HelmRelease refs for system charts
We are about to make potentially breaking changes to several of the gsp
systems as part of the migration to an eks control plane.

To avoid the systems getting rolled out to the "live" clusters we pin
them to the known good configurations.

This will prevent continuous deployment of these systems until we have
finished the migration.