Do not specify `labels` twice
With namespaces that we had specified `extra_namespace_configuration`
and that `namespace_configuration` included `labels` we would overwrite
the ones defined previously that added a label with the name of the
namespace.

Fortunately the `extra_namespace_configuration` was only ever used for
setting `labels` so we can easily change it.