provisions ingress eip per availability-zone
we are going to need static IP addrs for our ingress NLBs. this will
require one eip per az.

we don't yet have anything to attach it to as there is ongoing work to
provision the ingress bits, but we need to start telling people about
the addrs so are doing this a bit early.

I renamed the "nat_gateway_public_ip" inputs/outputs/resource since it
makes it a bit clearer that we have an ingress eip and egress eip for
each az (6 total).