Restart Kiam agent and server after deploy
Context: each time we deploy we re-render the Helm templates and in
doing so re-generate a new CA, keys and certs for Kiam.

Kiam server has an `exec` health check which tries to connect to itself
using generated keys, stored on the filesystem and populated by a
ConfigMap. After a deploy, when new keys have been generated, the keys
on disk used by this health check will not work.  This causes the
liveness check for the server to fail because the health check uses the
new keys and for the Kiam server uses the old keys. This causes the Kiam
server pods to restart. This causes the server to read in the new
CAs/keys/certs that won't match the Kiam agent.

To workaround we are forcing the Kiam pods to be rolled after each and
every deploy. We do this by forcing an environment variable change which
results in a Pod `spec` change which triggers the rolling update.