ingress nlb and target groups
create target groups for http (40080) and https (40443) traffic and
wire them to the autoscaling groups as part of the nodegroup
provisioning.

port 40080 for http and port 40443 were chosen arbitarily.

an NLB is provisioned that forwards traffic to the target groups:

 80 -> 40080
443 -> 40443

static ingress eips are passed through from the network module and
attached to the NLB , but maybe should be moved to the gsp-cluster
module where they are consumed.