Create istio ingress for the canary.
Updated the istio configuration (including a refactor of the istio chart
template) to allow ingress gateways to be created for the canary. External
DNS and cert manager are also used to provision the certificates and
route53 records for the wildcard ingress endpoints for the cluster.

Also fixed SSM deployment and added kiali to aid visualising and debugging
the istio mesh.