adds a cluster_number variable for picking netnum
we'd like to allow picking unique networks for each persistant cluster
so that we can avoid network clashes between clusters or other private
networks if we ever have to bridge them.

ideally each cluster would have a unique cluster_number (0-255).

it defaults to 0 so as not to disrupt existing deployments.