removes flux, helm-operater and tiler
now that we have concourse in each cluster and not just in "tools"
we are replacing flux/helm-operator/tiler with concourse to give us
(and our users) more visibility into the deployment flow, make it
easier to reason about, and give us less to manage.

this removes all mentions of flux/helm from the repo, some manual
intervention will be required to remove the CRDs and running deployments
from existing clusters.