Enable Cloudwatch Acceptance Tests
This enables cloudwatch acceptance tests on a cluster by discovering the latest log received by cloudwatch.

This test run from the central concourse against a cluster after deployment has occured.

Work was done to integrate this functionality within sonobuoy as a plugin, but was deemed unfeasible as the complexity of implementing and maintaining it would cause signicant issues going forward.  Testing so far has also shown sonobuoy to be somewhat flakey.

Reasons why sonobuoy was not chosen as the acceptance test technology:

1. Issues and complexity around passing secrets in to sonobuoy to allow certain tests to be run
1. The wrapping framework of sonobuoy makes it challenging to pass in arguments to direct tests
1. The learning curve of sonobuoy to implement tests is high
1. To build tests within sonobuoy, docker containers need to created and managed, which is a high cost for simple acceptance tests

However, sonobuoy could still be useful in the future for any tests that need to run within cluster and test the internal components.

Initial work explored the used of "aws logs describe-log-streams", however this did not produce reliable results as it requires prior knowledge of what logstreams within a loggroup should be monitored.  The logstreams appear to be buffered and can suffer delays when passing through fluentd which would sometimes incorrectly cause the test to fail.  Using "aws logs filter-log-events" instead avoided this problem.

Solo: @smford