enables Pipeline validation webhook
updates the concourse-operator to a version that includes a validating
webhook for pipeline CRDs

without validation we often see the operator stuck in a loop of
attempting to apply an invalid pipeline config.

there is no visibility into this failure other than checking the logs
(which is impractical for users without access to platform logs)

this should make it so that invalid pipelines fail at `kubectl apply`
time and so surface errors early.