less strict pipeline validation
using strict yaml unmarshalling is too strict for many pipelines and
adds little value to the validation process.

yaml anchors are often used on nonexistant/ignored properties as a way
of making pipelines with reapeatable chunkcs easier to read, this causes
validation to fail since it cannot unmarshal the yaml.

unmarshalling was also being considered an error for the webhook rather
than a validation failure, which was incorrect.