Fix the destroy pipeline
Now that we are not allowing accidental removal of the EKS cluster, the
pipeline is broken.

A way to solve it, is to clear out the config file and performing an
apply action on it against the same account. This should trigger
destroyment, and ignore any `prevent_destroy` settings previously set
up.