disables SNAT of Pod network traffic
We were having trouble adding protective monitoring that made us of the
VPC Flow logs as _sometimes_ traffic was getting routed to external
services also in the private VPC via the primary node interface(s)
instead of what we expected to see as coming directly from the Pod addr.

By default, the Amazon VPC CNI plugin for Kubernetes configures pods
with source network address translation (SNAT) enabled. This sets the
return address for a packet to the primary public IP of the instance and
allows for communication with the internet.

SNAT is only required for nodes that reside in a public subnet but our
nodes reside in private subnet and connect to the internet through a NAT
gateway so we should not need it.