Check if Canary has been deployed recently
At the moment, the canary could have been deployed once and then fail
thereafter. Kubernetes will keep it up despite it further attempts to
re-deploy it. This means we cannot rely on the ping only.

The `/metrics` endpoint exposes the timestamp of when the canary has
been deployed.

We can read that entry and attempt to find out if it hass been deployed
in the past N minute slot. That will assure us, that it has in fact been
re-deployed.