Re-use the secret for different components
When we deploy Harbor in our release pipeline on shared concourse, we're
running `helm template`. This causes some of the values to be generated.
The newly templated app is then kube cuttle applied to the system, and
perhaps does not roll all the boxes properly - as a version of a chart
hasn't really changed.

This is potentially causing a problem we are experiencing, where each
run of the release pipeline breaks the Registry. The solution often is
to kick the containers that haven't been restarted thus far and it will
sort itself out.

By hardcoding / re-using the secret, we hope for it to persist through
all deployments, making our lives easier and preventing the need to kick
the containers into good shape.