Improve access to CloudWatch
At the moment, each user is capable of getting onto the AWS account and
dig through the CloudWatch logs. This is however restricted to the CW
itself and does not include CloudWatch Insights.

We can replace manually created actions with a policy that is attached
to the user, that should give everyone readonly access to everything and
is managed by AWS themselves.