Allow all users to be Viewers
There is no reason for people to have a very limited access to the
resources used on AWS.

This role allows only for Describing and Listing different pieces of
infrastructure and should prevent things like s3:GetObject from
happening accidentally.

This policy is managed by AWS meaning we have to worry less (?).