Add cluster-autoscaler
We are using cluster-autoscaler 1.14.5 to match our kubernetes
version.  We'll have to remember to upgrade cluster-autoscaler when we
upgrade the cluster.

This has been adapted to work with the new per-AZ nodegroups.  For
this reason, we set `balance-similar-node-groups` so that the
cluster-autoscaler knows to try to keep the different AZ ASGs roughly
the same size within the constraints it has.

The cluster-autoscaler has the `system-cluster-critical` priority
because it should always be able to run.  If it were unable to run in
a busy cluster then the cluster could get stuck in a state where it
couldn't launch any new nodes.

We automatically detect which ASGs to scale based on the
`k8s.io/cluster-autoscaler/enabled` and
`k8s.io/cluster-autoscaler/${ClusterName}` tags.