disables exporters for eks managed components
EKS manages etcd, controller-manager and scheduler in a way that is not
observable via the usual kubernetes means, this results in false alerts
from prometheus which is configured to expect to find them in the
cluster. To prevent these false posatives we disable the
exporter/ServiceMonitors for the managed etcd, controller-manager, and
scheduler.