updates constraint to pull in concourse v5.3.0 lib
concourse-operator was being built with an old version of the
concourse/atc library that was missing some recent pipeline config
features, as a result pipeline validation was failing for pipelines with
things like in_parallel steps (concourse v5 feature)

this updates the dep depenedency constraints so that a newer version is
pulled in.

since we are not commiting the vendor dir right now, the Dockerfile also
required updating to perform `dep ensure` before a build.

some tests were added to ensure we don't regress on this issue.

resolves #231