Skip only merge commits
We'd like an empty commit created by the user not to be skipped during
our checks.

AFAIK, the merge commits will always have more than one parent whilst
the ordinary commit should only keep one. This is a safe assumption for
a check and to prevent an exclusion of empty commits by developers.