use fork of github-release resource
there is a bug in the github-release resource that causes it to only
fetch the first "page" of results from the github release (first 30
files) leading to confusing failures where files are missing.

To workaround 30 file limit in the meantime we use a forked build
that bumps the limit to the max allowed for a single page to 100.

There is a PR open upstream to resolve. Once upstream is fixed we should
stop using this fork.