make bash clearer
it's an oxymoron i know

I got confused by these lines - the release only has one file which
fits each of these wildcards - `source.tar.gz.asc` and `source.tar.gz`
respectively.  I found it confusing that these were wildcards because
it looks like it's trying to validate many signatures and unpack many
tarballs (and there are other tarballs in the release, only with `tgz`
in the name instead of `tar.gz`).

I think it's much clearer to use the literal filenames instead of a
wildcard.