Remove path from namespace roles
We found that the path is actually included in the resource ARN.
That's fine until you consider that to call the STS AssumeRole API, you need
to provide that full ARN - and so we'd need to put the path in our kubeyaml
namespace permitted roles regexes as well as pod definitions.

The path doesn't really do anything anyway so let's just get rid of it.