Define deployment chart for ServiceOperator
We'd like to deploy the service operator as part of the cluster
deployment/release pipeline.

In order to do that, we are re-using the release pipeline to build the
thing, which also coincidentally packages each component into the
system.

Building the component is not enough, it will need to have the
definition of KubeYAML that deploys the app and CRD into place.

Co-authored-by: Issy Long <isabell.long@digital.cabinet-office.gov.uk>
Co-authored-by: Rafal Proszowski <rafal.proszowski@digital.cabinet-office.gov.uk>