Add CRD fields and Operator options
We came to realisation, that the current fields may not be enough.

We will need a cluster name, in order to pin point any service to a
specific cluster.

We are adding a `secret` field into the spec, so that a user can name
what the secret would be. We should then be falling back into a default
name later if one is not specified.

Our status, is gaining a Reason as well as ID of an instance, for
further reference if there would be a need of debugging.

Co-authored-by: Alex Monk <alex.monk@digital.cabinet-office.gov.uk>
Co-authored-by: Issy Long <isabell.long@digital.cabinet-office.gov.uk>
Co-authored-by: Rafal Proszowski <rafal.proszowski@digital.cabinet-office.gov.uk>