Add support for multi-group
It is possible to have multi-groups with the new version of KubeBuilder,
however it's not yet available in beta with the use of generator. This
can however be fixed manually with not that much effort by following the
[documentation][1] from the KubeBuilder book.

This should make a nice separation of services we would like to provide
with this operator and allow it to be a little more generic.

[1]: https://book.kubebuilder.io/migration/multi-group.html