Scrape kubelet metrics with node-role label
I would like us to scrape kubelet metrics in such a way that we can
tell which node has which role.  This is so that, for example, I can
see if a particular node role is running out of resources.  (eg: if
there are no pods available in the worker node group, it's no good
knowing that the CI or cluster-management nodes have plenty of pods
available.)

Currently, this isn't possible for tiresome reasons.

 1. Node roles are set as node labels (node-role.kubernetes.io/<role>)
 2. The labels have an empty value
 3. Prometheus doesn't distinguish between missing labels and labels
    with an empty value

This means that, although it's possible to see node labels via the
`kube_node_labels` metric, the node role labels are not there.

Prometheus 2.9.0 introduced a workaround: you can see if node labels
are present using the `__meta_kubernetes_node_labelpresent_<label>`
meta label in your `kubernetes_sd_configs`.  These labels are only
present if you are doing node-based service discovery (ie using `role:
node`).

Even more tiresomely, the prometheus operator defaults to scraping
kubelet via endpoint-based service discovery, for all sorts of reasons
which need not concern us here.

Therefore, I have decided to re-scrape kubelet via
`additionalScrapeConfigs` that specify `role: node` and include checks
for the specific node-role labels.  This way, we can get a `node_role`
target label on our kubelet metrics, and finally, once and for all,
actually have metrics that can be sliced by node group.

:weary:

By the way, this is untested.  If it works, we should switch off the
old kubelet metrics via the following in gsp-cluster/values.yaml:

    prometheus-operator:
      ...
      # disable service/gsp-prometheus-operator-kubelet in kube-system
      prometheusOperator:
        kubeletService:
          enabled: false
      # disable the kubelet ServiceMonitor that scrapes the above service
      kubelet:
        enabled: false