Automate chart generation
At the moment, we're having essentially a tmp state of generated
resources. Then these are being concated into a file in our charts
directory.

Role resource is not capable of being autogenerated due to the kustomize
not understanding helm's templating. This way, we're loosing some magic
permission generation, meaning we have to manually keep the resource
in right shape to allow operator operate.

Little piece of sed can help with that.