execute unit tests as part of docker build
run the unit tests as part of the build process to ensure that images
cannot be built if they have failing tests and also to make it easier to
execute the tests without having to configure your developerment
environment.

the tests require the version of kubebuilder that was used to create the
controllers (along with binaries for the controlplane) so these
dependencies are added to the build container.