Attempt to fix commit signing check
We tell the PR resource to use checkout during the initial get, but then we put
pending status. When you put, it does an implicit get, which defaults to merge.
Then in our code we see the merged copy, which will not be a commit in the
upstream GitHub repo, so the check fails.