seperate make task for update-chart
we need the manifest aritifacts generated for the service-operator
project itself (tests, etc) and not only for the parent chart.

this keeps the default kubebuilder makefile behaviour of generating
artifacts in config dir but provides an update-chart task for helping
update the parent gsp-cluster chart

the tests expect the generated artifacts to exist, so generate them
during build for tests.