WIP: restrict HSM access to specific namespaces
This adds a rule to our GlobalNetworkPolicy that restricts access to
the HSM to specific, named namespaces.

The idea is: currently, it's technically possible for a tenant in a
non-proxy-node-managed namespace to open up access to the HSM for
themselves, by setting up appropriate NetworkPolicies and
ServiceEntries and DestinationRules and what have you, and by setting
the `talksToHsm` label to `true` on a particular pod.  There are a
bunch of controls to stop them doing it - see ADR036 for details - but
it remains a possibility.

If we add a GlobalNetworkPolicy rule that specificially identifies the
namespaces which are allowed to talk to the HSM, that completely shuts
down any possibility for somebody to open a network connection to the
HSM from an unauthorised namespace.  This feels like a win.

This should not be merged without an ADR amending ADR036 with the
updated network control.

A possible future improvement would be to set a `talksToHsm` label on
the appropriate namespaces, rather than having Verify-specific
namespace names in the global gsp repository.  (It would then be up to
namespace.yaml and/or verify-cluster-config to apply this label to the
appropriate namespaces).  Tenants should not be able to set labels on
their own namespaces, so this would provide the same level of security
control, without coupling the Verify-specific config to the global
cluster config.

However, although that would be a nicer design (IMHO), it would be
more work, and simply hardcoding the specific namespaces is better
than nothing.  We could always migrate to that design in future.