Add sentry to the GSP offering
Sentry is a nice system allowing users to throw the exceptions at to be
stored, reviewable and alerted upon.

This particular solution is used by Verify, GOV.UK, GovWifi, soon by Pay
and Notify.

Making it a part of GSP could help in keeping the stack in a single box.