Provision a Principal per-namespace
Trying to provision a Principal at the same time as any other AWS
resources using the service operator can lead to a race condition where
the IAM role hasn't been created by the time CloudFormation tries to
provision the other resources. At this point the stack creation with
fail and everything ends up in a bad state.

We might as well provision a Principal when setting up the namespace
to allow for trouble-free provisioning of AWS services in the
namespace's own charts.