Update `calico.yaml`
This `calico.yaml` manifest is provided by AWS[1,2]. This change updates
to the latest manifest provided by AWS which makes some changes but
doesn't bump the Calico version.

There are unreleased versions of this manifest[3] which upgrade to newer
versions of Calico but we should probably not run anything other than
officially released versions.

This change also includes the removal of the Calico autoscaler which we
did for performance reasons in
`c60f78d2bde4e991da8b98c1fc5bf34ed1d26323`. We should look at if this
still a sensible thing to do later.

[1] https://docs.aws.amazon.com/eks/latest/userguide/calico.html
[2] https://raw.githubusercontent.com/aws/amazon-vpc-cni-k8s/release-1.5/config/v1.5/calico.yaml
[3] https://github.com/aws/amazon-vpc-cni-k8s/blob/master/config/v1.5/calico.yaml