allow SREs to clean up prometheus config
we have some stray `servicemonitors` that we want to clean up.
Specifically, since 583ee5053 we no longer need
servicemonitor/gsp-prometheus-operator-kubelet in gsp-system and we'd
like to delete it.

I think we should be able to delete any of the prometheus config
types, so I've added them all as things that SREs can delete.