clarify optional PolicyAttacher/SecretOutputter
implmenting StackPolicyAttacher (to request a role name to attach a
policy to) or StackSecretNamer (to name a Secret that should be
populated with cloudformation outputs) is optional. The code was not
making this clear previously as the casts were being attetmpted at a
point that made them look like unchecked error conditions. By checking
the codition before the call the intension is hopefully clarified.