relax concourse github PR approval rules
Currently, our concourse-github-resource requires that the author of a
PR is the same person as the signer of the last commit in the PR.
This is annoying when someone opens a PR and then goes on leave or is
ill or something and somebody else wants to finish off the work.

This changes the behaviour so that we simply discount the commit
signer from being able to approve their own work.  If you sign a
commit and push to a PR (regardless of who originally opened the PR),
we ignore your approval and you need to get someone else to approve
it.

This is more permissive than the previous version, but it still
enforces that nobody can approve their own code.

I tested this by running locally against two PRs:

 - #495 - opened by @samcrang, re-signed by me, approved by
   @blairboy362 and @Krenair; this was approved (but would be rejected
   by the current resource)
 - #496 - opened by @samcrang, re-signed by me and approved by me;
   this was rejected (because I shouldn't be able to approve my own
   commit)

@samcrang saw me run these manual tests.