[release pipeline] improve package-deployer-vars error handling
the templating here is broken but the pipeline didn't fail.  We need
the bash "cause pipeline to fail" setting aka `pipefail` :troll:

(More seriously, `pipefail` causes commands to error if a part of a
pipeline errors.  It seems that the `spruce` command failed but
because we pipe the output to `tee` the error was being swallowed.)