create deployer.defaults.yaml fresh, don't append
The file we were creating had a bunch of duplicate fields.  What seems
to be happening is that deployer.defaults.yaml exists already in the
target location, and we append to it, with a bunch of duplicate keys.

We should just create the whole file from scratch here i think?