[concourse-github-resource] require github-verified signatures
We were fetching `isValid` in the graphql query but not using it for
anything.  I think it's probably useful to require that github has
verified the signature, because that means that we can be confident
that github thinks the signing key actually belongs to the specified
user account.

Without this, I can vaguely imagine an attack where:

 - I add a new key for myself to gds-trusted-developers
 - I *don't* add the key to my github account
 - (I *also* ensure I don't use an email address for a key userid that
   matches any of my github email addresses)
 - the signature is approved because the key is in gds-trusted-developers
 - github can't match the signature to my github user, so the graphql
   query doesn't know that I am the signer
 - I am therefore allowed to approve my commit because github doesn't
   realise it's mine
 - i have skipped the two-eyes process

With this check, the attack won't work because we refuse to allow a
signed commit where github can't match the signature to a user.

This is a pretty academic attack (it leaves an awful lot of paper
trail behind) but I think it's enough to make this check worth doing.