expands allowed actions in provisioned SQS policy
Concumers of SQS queues need more permissions than were previously
granted. This expands the roles to allow all the queue-scopeable actions
that may be required and do not conflict with actions the operator will
take on the stack (ie SetQueueAttributes would conflict with
cloudformation config so is disallowed).

We have seen issues where ListQueues is required to fetch some queue
infomation. This is unfortunatly not a "queue-scopable" action and as a
result this permission will give access to view ALL existing queues.
This is far from ideal, but the data that leaked is the
queue name, and attributes not access to the queue itself.