Deploy 2 instances of Prometheus
Prometheus's fault-tolerance is based on it being a really dumb simple
thing to operate.  It doesn't do replication, backups, leadership
elections, distributed locking, or anything like that.  It's just a
single instance with a local disk.

As a result, if we lose the AZ that our single Prometheus is running
in, we'll lose Prometheus.  To guard against that, we should maybe
have 2 of them?

In a multiple Prometheus setup, each Prometheus is completely
independent.  Each Prometheus will scrape each target and maintain a
separate tsdb of time series data.  As a result, it's probably best to
always target the same Prometheus instance where possible so that you
don't get weird jitter in your results where different Prometheis
scraped out of phase with each other.

Therefore, this also sets the Prometheus service session affinity to
`ClientIP` which enforces sticky sessions. See
https://kubernetes.io/docs/concepts/services-networking/service/#virtual-ips-and-service-proxies
for more on this.

I considered making this configurable on a per-cluster basis but I
thought that was probably more effort than it was worth.  2 is a
sensible default.